Add files via upload
1. Write a Java program that uses polymorphism by defining an interface called Shape with methods to calculate the area and perimeter of a shape. 
Then create classes that implement the Shape interface for different types of shapes, such as circles and triangles.
2. Write a Java program to invoke parent class constructor from a child class. Create Child class object and parent class constructor must be invoked. 
Demonstrate by writing a program. Also explain key points about Constructor.
3. Write a Java programme that takes an integer from the user and throws an exception if it is negative.
Demonstrate Exception handling of same program as solution.
4. Create a Java program that simulates a bank account. The program should allow users to deposit and withdraw money, check their balance.
5. Demonstrate the difference between abstract class and interface by writing programs as well as in keypoints.
6. Write a Java program that uses stream api to perform operations on a large data set, such as sorting or filtering the data.
7. Create a Java program that implements a binary search algorithm. The program should accept user input for the target value and search for it in a sorted array. 
The program should return the index of the target value if found or a message if not found.
8. Write a Java program that creates two threads. The first thread should print even numbers between 1 and 10,
 and the second thread should print odd numbers between 1 and 10.
9. Write a Java program that implements a producer-consumer model using multithreading. The program should have a producer thread that generates random
numbers and adds them to a queue, and a consumer thread that reads numbers from the queue and calculates their sum. The program should use synchronization to
ensure that the queue is accessed by only one thread at a time.
10. Write a Java program that reads a set of integers from the user and stores them in a List. 
The program should then find the second largest and second smallest elements in the List.