Question 24
Create a Spring Boot application that uses Spring MVC to create a REST API. The
API should accept a JSON request with data and insert it into a MySQL database
table using JPA and Hibernate. The application should use Spring Data JPA to map
the table to a Java object and then insert the data into the table.