Question 17
 Create a web application that lets users create and view blog posts. The web
application should use the MVC pattern, with servlets as controllers, JSPs as views,
and a database as the model. Users should be able to create new blog posts by
filling out a form that includes a title, description, and content. The web application
should use a servlet to store the blog post data in the database. Users should also be
able to view all the blog posts on a separate page, and the web application should
use a servlet to retrieve the blog post data from the database and display it in a
formatted way